// types/CryptoCurrency.ts
export type CryptoCurrency = {
    id: string;
    market_cap_rank: number;
    symbol: string;
    name: string;
    image: string;
    current_price: number;
    price_change_percentage_1h_in_currency: -0.15320204482591082;
    price_change_percentage_24h_in_currency: -2.104755096160124;
    price_change_percentage_7d_in_currency: 0.8674673760884438;
    market_cap: number;
    circulating_supply: number;
    total_volume: number;
    sparkline_in_7d: number[];
  };